Preserve float timestamps in safe_dump cast; add at-rest detector
Two follow-ups to the #3424 boundary cast:

1. created/updated now cast with to_f, not to_i. These timestamps are
   float epoch seconds and double as sorted-set scores (instances
   timeline, expiration timeline); truncating to whole seconds would
   reorder range queries and discard valid precision. The V3 contract is
   z.number(), which accepts either, so there is no reason to narrow it.
   TTL/lifespan fields keep to_i since they are integer-second durations
   with no sub-second component to lose.

2. Add scripts/diagnostics/detect_string_typed_numerics.rb, a read-only
   scan for records already poisoned at rest. The cast fixes the wire
   format but leaves the stored bytes corrupt; this finds them. It is the
   mirror image of check_raw_email_fields.rb (#3016): that flags non-JSON
   bytes, this flags valid JSON whose parsed type is String where a number
   is required. The report includes a per-record signature (whether a
   later save healed `updated` while other fields stayed strings) to help
   trace which writer produced the bad data. The auto-run guard is keyed
   on $PROGRAM_NAME only, so requiring the file (e.g. from the tryout that
   now exercises its predicate) never triggers a keyspace scan.

Mechanism tryout updated: created/updated assertions are Float again, and
a new section pins the detector predicate (quoted numeric -> flagged,
bare number/float/nil/non-JSON -> not flagged) plus an end-to-end check
against real poisoned and healthy hgetall output.

Verified: tryouts 28 + receipt_safe_dump 23 + safe_dump 4; rspec receipt
+ v1/v2 secrets 198 examples 0 failures; vitest store + contracts 122.

+++ lib/onetime/models/receipt/features/safe_dump_fields.rb
@@ -63,20 +63,20 @@ module Onetime::Receipt::Features
         }
       base.safe_dump_field :secret_identifier
 
-      # Numeric fields are cast at this boundary because Familia v2 storage
-      # is type-preserving, not type-enforcing: a String written upstream
-      # hydrates as a String and fails the strict z.number() V3 schema. The
-      # cast is a no-op for healthy records; unset secret_ttl stays -1 and
-      # unset lifespans stay nil, as before. See the longer note in
-      # secret/features/safe_dump_fields.rb and #3424.
-      # Mechanism tests: try/unit/models/secret_numeric_field_types_try.rb
+      # Cast numeric fields at the serialization boundary — see the full note
+      # in secret/features/safe_dump_fields.rb (#3424). TTL/lifespan use to_i
+      # (lossless integer-second durations; secret_ttl keeps -1 and lifespan
+      # keeps nil when unset). created/updated use to_f, NOT to_i, to preserve
+      # the sub-second precision that matters when these values are used as
+      # sorted-set scores. Detector:
+      # scripts/diagnostics/detect_string_typed_numerics.rb
       base.safe_dump_field :secret_ttl, ->(m) { m.secret_ttl.to_i > 0 ? m.secret_ttl.to_i : -1 }
       base.safe_dump_field :metadata_ttl, ->(m) { m.lifespan.to_i > 0 ? m.lifespan.to_i : nil }
       base.safe_dump_field :receipt_ttl, ->(m) { m.lifespan.to_i > 0 ? m.lifespan.to_i : nil }
       base.safe_dump_field :lifespan, ->(m) { m.lifespan.to_i > 0 ? m.lifespan.to_i : nil }
       base.safe_dump_field :share_domain
-      base.safe_dump_field :created, ->(m) { m.created&.to_i }
-      base.safe_dump_field :updated, ->(m) { m.updated&.to_i }
+      base.safe_dump_field :created, ->(m) { m.created&.to_f }
+      base.safe_dump_field :updated, ->(m) { m.updated&.to_f }
       base.safe_dump_field :shared
       # Obscure recipient emails at serialization time so the raw address
       # never reaches the frontend, while the underlying record keeps the
